chore: refactor validationSchema to helpers
