+++ packages/frontend/src/components/FlowStep/components/TestAgainInfobox.tsx
@@ -3,9 +3,9 @@ import { BiSolidErrorCircle } from 'react-icons/bi'
 import { Flex, Icon, Text } from '@chakra-ui/react'
 
 import { EditorContext } from '@/contexts/Editor'
+import { getFlowStepHeaderWidth } from '@/helpers/editor'
 
 import { flowStepStyles } from '../styles'
-import { getFlowStepHeaderWidth } from '../utils'
 
 interface TestAgainInfoboxProps {
   isNested?: boolean
